Add files via upload
A glass morphism is a glassy effect that is used in modern websites to provide a unique user interface. In the glassmorphism effect, we blur the background and highlight the content. HTML structures the form, and CSS styles it to achieve the glassy effect with blurred backgrounds and transparency. This design not only looks visually appealing but also enhances the user experience by creating an intuitive and attractive login interface, making it a popular choice for modern web designers.